# Conflicts:
#	SO3Fun/@SO3Fun/SO3Fun.m
#	SO3Fun/@SO3Fun/plotSection.m
#	SO3Fun/@SO3FunHarmonic/SO3FunHarmonic.m
#	SO3Fun/@SO3FunHarmonic/display.m
#	SO3Fun/@SO3FunHarmonic/eval.m

+++ TensorAnalysis/@complianceTensor/PoissonRatio.m
@@ -39,7 +39,7 @@ elseif nargin <= 2 || isempty(y)
 else
 
   % compute tensor product
-  nu = -double(EinsteinSum(S,[-1 -2 -3 -4],x,-1,x,-2,y,-3,y,-4)) ./ ...
-    double(EinsteinSum(S,[-1 -2 -3 -4],x,-1,x,-2,x,-3,x,-4));
+  nu = -EinsteinSum(S,[-1 -2 -3 -4],x,-1,x,-2,y,-3,y,-4) ./ ...
+    EinsteinSum(S,[-1 -2 -3 -4],x,-1,x,-2,x,-3,x,-4);
   
 end
